+++ CHANGELOG.md
@@ -9,7 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 ### Added
+- feat: Add `babelRegister` and `babelRegisterOptions` low level configuration options, allowing to enable Babel when loading mocks and routes (#151);
+- feat: Add `lowLevelConfig` getter to core, returning low level configuration properties.
 ### Changed
+- chore: update node versions in pipeline
 ### Fixed
 ### Removed
 ### BREAKING CHANGES